+++ typademic/errors/__init__.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
 
-bp = Blueprint('errors', __name__)
+blueprint = Blueprint('errors', __name__, template_folder='templates')
 
 from typademic.errors import handlers
