+++ typademic/templates/index.html
@@ -20,7 +20,7 @@
                 <div class="tile">
                     <div class="content">
                         <p class="circle">3</p>
-                        {{ dropzone.create(action='upload') }}
+                        {{ dropzone.create(action='/upload') }}
                         {{ dropzone.load_js() }}
                         {{ dropzone.config() }}
                     </div>
