+++ typademic/uploads/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+blueprint = Blueprint('uploads', __name__, template_folder='templates')
+
+from typademic.uploads import routes
